+++ modules/apps/web-experience/asset/asset-publisher-layout-prototype/src/main/java/com/liferay/asset/publisher/layout/prototype/internal/instance/lifecycle/AddLayoutPrototypePortalInstanceLifecycleListener.java
@@ -44,6 +44,11 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
+ * Provides the implementation of <code>PortalInstanceLifecycleListener</code>
+ * (in the <code>com.liferay.portal.instance.lifecycle</code> module), which
+ * adds a default layout prototype for the content display page when a new
+ * portal instance is registered.
+ *
  * @author Juergen Kappler
  */
 @Component(immediate = true, service = PortalInstanceLifecycleListener.class)