+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/display/context/AssetPublisherDisplayContext.java
@@ -75,6 +75,9 @@ import javax.portlet.PortletURL;
 import javax.servlet.http.HttpServletRequest;
 
 /**
+ * Provides utility methods moved from the Asset Publisher portlet's JSP files
+ * to reduce the complexity of the views.
+ *
  * @author Eudaldo Alonso
  */
 public class AssetPublisherDisplayContext {