+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/BaseSourceProcessor.java
@@ -516,14 +516,16 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		if (_portalLanguageProperties == null) {
 			Properties portalLanguageProperties = new Properties();
 
-			File portalLanguagePropertiesFile = new File(
-				getFile("portal-impl", PORTAL_MAX_DIR_LEVEL),
-				"src/content/Language.properties");
+			File portalLanguagePropertiesFile = getFile(
+				"portal-impl/src/content/Language.properties",
+				PORTAL_MAX_DIR_LEVEL);
 
-			InputStream inputStream = new FileInputStream(
-				portalLanguagePropertiesFile);
+			if (portalLanguagePropertiesFile != null) {
+				InputStream inputStream = new FileInputStream(
+					portalLanguagePropertiesFile);
 
-			portalLanguageProperties.load(inputStream);
+				portalLanguageProperties.load(inputStream);
+			}
 
 			_portalLanguageProperties = portalLanguageProperties;
 		}
@@ -567,8 +569,10 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 					continue;
 				}
 
+				BNDSettings bndSettings = getBNDSettings(fileName);
+
 				Properties bndFileLanguageProperties =
-					getBNDFileLanguageProperties(fileName);
+					bndSettings.getLanguageProperties();
 
 				if ((bndFileLanguageProperties != null) &&
 					!bndFileLanguageProperties.containsKey(languageKey)) {
@@ -576,6 +580,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 					processMessage(
 						fileName, "Missing language key '" + languageKey + "'");
 				}
+
+				putBNDSettings(bndSettings);
 			}
 		}
 	}
@@ -1702,70 +1708,15 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return _annotationsExclusions;
 	}
 
-	protected Properties getBNDFileLanguageProperties(String fileName)
-		throws Exception {
+	protected BNDSettings getBNDSettings(String fileName) throws Exception {
+		for (Map.Entry<String, BNDSettings> entry :
+				_bndSettingsMap.entrySet()) {
 
-		Tuple bndFileLocationAndContentTuple =
-			getBNDFileLocationAndContentTuple(fileName);
+			String bndFileLocation = entry.getKey();
 
-		if (bndFileLocationAndContentTuple == null) {
-			return new Properties();
-		}
-
-		String bndFileLocation =
-			(String)bndFileLocationAndContentTuple.getObject(0);
-
-		Properties properties = _bndLanguagePropertiesMap.get(bndFileLocation);
-
-		if (properties != null) {
-			return properties;
-		}
-
-		String bndContent = (String)bndFileLocationAndContentTuple.getObject(1);
-
-		if (bndContent.matches(
-				"[\\s\\S]*Provide-Capability:.*liferay\\.resource\\.bundle" +
-					"[\\s\\S]*")) {
-
-			// Return null, in order to skip checking for language keys for
-			// modules that use LanguageExtender. No fix in place for this right
-			// now.
-
-			return null;
-		}
-
-		Matcher matcher = bndContentDirPattern.matcher(bndContent);
-
-		if (matcher.find()) {
-			File file = new File(
-				bndFileLocation + matcher.group(1) + "/Language.properties");
-
-			if (!file.exists()) {
-				return new Properties();
+			if (fileName.startsWith(bndFileLocation)) {
+				return entry.getValue();
 			}
-
-			properties = new Properties();
-
-			InputStream inputStream = new FileInputStream(file);
-
-			properties.load(inputStream);
-
-			_bndLanguagePropertiesMap.put(bndFileLocation, properties);
-
-			return properties;
-		}
-
-		return new Properties();
-	}
-
-	protected Tuple getBNDFileLocationAndContentTuple(String fileName)
-		throws Exception {
-
-		Tuple bndFileLocationAndContentTuple =
-			_bndFileLocationAndContentMap.get(fileName);
-
-		if (bndFileLocationAndContentTuple != null) {
-			return bndFileLocationAndContentTuple;
 		}
 
 		String bndFileLocation = fileName;
@@ -1782,20 +1733,16 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			File file = new File(bndFileLocation + "bnd.bnd");
 
 			if (file.exists()) {
-				String bndContent = FileUtil.read(file);
-
-				bndFileLocationAndContentTuple = new Tuple(
-					bndFileLocation, bndContent);
-
-				_bndFileLocationAndContentMap.put(
-					fileName, bndFileLocationAndContentTuple);
-
-				return bndFileLocationAndContentTuple;
+				return new BNDSettings(bndFileLocation, FileUtil.read(file));
 			}
 
 			bndFileLocation = StringUtil.replaceLast(
 				bndFileLocation, StringPool.SLASH, StringPool.BLANK);
 		}
+	}
+
+	protected Map<String, BNDSettings> getBNDSettingsMap() {
+		return _bndSettingsMap;
 	}
 
 	protected Map<String, String> getCompatClassNamesMap() throws Exception {
@@ -2161,23 +2108,15 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			releaseVersion = ReleaseInfo.getVersion();
 		}
 		else {
-			Tuple bndFileLocationAndContentTuple =
-				getBNDFileLocationAndContentTuple(fileName);
+			BNDSettings bndSettings = getBNDSettings(fileName);
 
-			if (bndFileLocationAndContentTuple == null) {
+			releaseVersion = bndSettings.getReleaseVersion();
+
+			if (releaseVersion == null) {
 				return null;
 			}
 
-			String bndContent =
-				(String)bndFileLocationAndContentTuple.getObject(1);
-
-			Matcher matcher = bndReleaseVersionPattern.matcher(bndContent);
-
-			if (!matcher.find()) {
-				return null;
-			}
-
-			releaseVersion = matcher.group(1);
+			putBNDSettings(bndSettings);
 		}
 
 		int pos = releaseVersion.lastIndexOf(CharPool.PERIOD);
@@ -2722,6 +2661,10 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		_modifiedFileNames.add(file.getAbsolutePath());
 	}
 
+	protected void putBNDSettings(BNDSettings bndSettings) {
+		_bndSettingsMap.put(bndSettings.getFileLocation(), bndSettings);
+	}
+
 	protected Document readXML(String content) throws DocumentException {
 		SAXReader saxReader = SAXReaderFactory.getSAXReader(null, false, false);
 
@@ -2793,7 +2736,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return line;
 	}
 
-	protected String sortMethodCalls(
+	protected String sortMethodCall(
 		String content, String methodName, String... variableTypeRegexStrings) {
 
 		Pattern codeBlockPattern = Pattern.compile(
@@ -2848,6 +2791,22 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				previousPutOrSetParameterName = putOrSetParameterName;
 			}
 		}
+
+		return content;
+	}
+
+	protected String sortMethodCalls(String absolutePath, String content) {
+		if (isExcludedPath(METHOD_CALL_SORT_EXCLUDES, absolutePath)) {
+			return content;
+		}
+
+		content = sortMethodCall(
+			content, "add", "ConcurrentSkipListSet<.*>", "HashSet<.*>",
+			"TreeSet<.*>");
+		content = sortMethodCall(
+			content, "put", "ConcurrentHashMap<.*>", "HashMap<.*>",
+			"JSONObject", "TreeMap<.*>");
+		content = sortMethodCall(content, "setAttribute");
 
 		return content;
 	}
@@ -2980,6 +2939,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return line;
 	}
 
+	protected static final String METHOD_CALL_SORT_EXCLUDES =
+		"method.call.sort.excludes";
+
 	protected static final String RUN_OUTSIDE_PORTAL_EXCLUDES =
 		"run.outside.portal.excludes";
 
@@ -2988,10 +2950,6 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		Pattern.MULTILINE);
 	protected static Pattern attributeNamePattern = Pattern.compile(
 		"[a-z]+[-_a-zA-Z0-9]*");
-	protected static Pattern bndContentDirPattern = Pattern.compile(
-		"\\scontent=(.*?)(,\\\\|\n|$)");
-	protected static Pattern bndReleaseVersionPattern = Pattern.compile(
-		"Bundle-Version: (.*)\n");
 	protected static Pattern emptyArrayPattern = Pattern.compile(
 		"((\\[\\])+) \\{\\}");
 	protected static Pattern emptyCollectionPattern = Pattern.compile(
@@ -3084,9 +3042,29 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		if (getFile("portal-impl", PORTAL_MAX_DIR_LEVEL) != null) {
 			return true;
 		}
-		else {
-			return false;
+
+		// Subrepositories should be treated as portal
+
+		String baseDirAbsolutePath = getAbsolutePath(
+			sourceFormatterArgs.getBaseDirName());
+
+		int x = baseDirAbsolutePath.length();
+
+		for (int i = 0; i < 2; i++) {
+			x = baseDirAbsolutePath.lastIndexOf(CharPool.FORWARD_SLASH, x - 1);
+
+			if (x == -1) {
+				return false;
+			}
+
+			String dirName = baseDirAbsolutePath.substring(x + 1);
+
+			if (dirName.startsWith("com-liferay-")) {
+				return true;
+			}
 		}
+
+		return false;
 	}
 
 	private String _normalizePattern(String originalPattern) {
@@ -3103,10 +3081,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	}
 
 	private Set<String> _annotationsExclusions;
-	private final Map<String, Tuple> _bndFileLocationAndContentMap =
-		new HashMap<>();
-	private final Map<String, Properties> _bndLanguagePropertiesMap =
-		new HashMap<>();
+	private Map<String, BNDSettings> _bndSettingsMap =
+		new ConcurrentHashMap<>();
 	private Map<String, String> _compatClassNamesMap;
 	private String _copyright;
 	private final Pattern _definitionPattern = Pattern.compile(
@@ -3165,6 +3141,17 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 			if (matcher.find()) {
 				putOrSetParameterName2 = matcher.replaceAll(StringPool.BLANK);
+			}
+
+			if (putOrSetParameterName1.matches("\".*\"") &&
+				putOrSetParameterName2.matches("\".*\"")) {
+
+				String strippedQuotes1 = putOrSetParameterName1.substring(
+					1, putOrSetParameterName1.length() - 1);
+				String strippedQuotes2 = putOrSetParameterName2.substring(
+					1, putOrSetParameterName2.length() - 1);
+
+				return super.compare(strippedQuotes1, strippedQuotes2);
 			}
 
 			int value = super.compare(