+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/JSPSourceProcessor.java
@@ -385,6 +385,9 @@ public class JSPSourceProcessor extends BaseSourceProcessor {
 			}
 		}
 
+		newContent = newContent.replaceAll(
+			"(\\s(page|taglib))\\s+((import|uri)=)", "$1 $3");
+
 		if (_stripJSPImports && !_jspContents.isEmpty()) {
 			try {
 				newContent = formatJSPImportsOrTaglibs(
@@ -456,9 +459,7 @@ public class JSPSourceProcessor extends BaseSourceProcessor {
 
 		checkSubnames(fileName, newContent);
 
-		newContent = sortMethodCalls(
-			newContent, "put", "HashMap<.*>", "JSONObject");
-		newContent = sortMethodCalls(newContent, "setAttribute");
+		newContent = sortMethodCalls(absolutePath, newContent);
 
 		newContent = formatStringBundler(fileName, newContent, -1);
 
@@ -835,9 +836,7 @@ public class JSPSourceProcessor extends BaseSourceProcessor {
 					line = StringUtil.replace(line, "<%=", "<%= ");
 				}
 
-				if (javaSource &&
-					trimmedLine.matches("^\\} ?(catch|else|finally) .*")) {
-
+				if (trimmedLine.matches("^\\} ?(catch|else|finally) .*")) {
 					processMessage(
 						fileName, "There should be a line break after '}'",
 						lineCount);
@@ -2136,7 +2135,7 @@ public class JSPSourceProcessor extends BaseSourceProcessor {
 		"jsp.unused.variables.excludes";
 
 	private final Pattern _compressedJSPImportPattern = Pattern.compile(
-		"(<.*\n*page.import=\".*>\n*)+", Pattern.MULTILINE);
+		"(<.*\n*page import=\".*>\n*)+", Pattern.MULTILINE);
 	private final Pattern _compressedJSPTaglibPattern = Pattern.compile(
 		"(<.*\n*taglib uri=\".*>\n*)+", Pattern.MULTILINE);
 	private final Pattern _defineObjectsPattern = Pattern.compile(
@@ -2201,7 +2200,7 @@ public class JSPSourceProcessor extends BaseSourceProcessor {
 	private final Pattern _testTagPattern = Pattern.compile(
 		"^<c:(if|when) test=['\"]<%= (.+) %>['\"]>$");
 	private final Pattern _uncompressedJSPImportPattern = Pattern.compile(
-		"(<.*page.import=\".*>\n*)+", Pattern.MULTILINE);
+		"(<.*page import=\".*>\n*)+", Pattern.MULTILINE);
 	private final Pattern _uncompressedJSPTaglibPattern = Pattern.compile(
 		"(<.*taglib uri=\".*>\n*)+", Pattern.MULTILINE);
 	private String _utilTaglibSrcDirName;