+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -714,11 +714,52 @@ moved to [Liferay Gradle Plugins].
 ### Changed
 - [LPS-69492]: Update the [Liferay Gradle Plugins] dependency to version 3.0.45.
 
+## 2.1.0 - 2016-12-03
+
+### Added
+- [LPS-68289]: Add module agent support for aspect in the `test` and
+`testIntegration` tasks.
+
+### Changed
+- [LPS-69518]: Update the [Liferay Gradle Plugins] dependency to version 3.0.46.
+
+## 2.1.1 - 2016-12-05
+
+### Changed
+- [LPS-69501]: Update the [Liferay Gradle Plugins] dependency to version 3.0.47.
+
+## 2.2.0 - 2016-12-08
+
+### Added
+- [LPS-63943]: Add the ability to create additional *prep next* commits via
+`PrintArtifactPublishCommands` tasks.
+
+### Changed
+- [LPS-63943]: Commit themes' `.digest` files in a separate *prep next* commit.
+- [LPS-69618]: Update the [Liferay Gradle Plugins] dependency to version 3.0.48.
+
+## 2.2.1 - 2016-12-08
+
+### Changed
+- [LPS-69501]: Update the [Liferay Gradle Plugins] dependency to version 3.0.49.
+
+## 2.2.2 - 2016-12-12
+
+### Added
+- [LPS-69606]: Publish the WSDD fragment JAR of an OSGi project with the
+`install` and `uploadArchives` tasks.
+
+### Changed
+- [LPS-69501]: Use the [Liferay Source Formatter] JAR files deployed locally in
+the `tools/sdk/dependencies/com.liferay.source.formatter/lib` directory, if
+found.
+
 [Liferay CDN]: https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
 [Liferay Gradle Plugins App Javadoc Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-app-javadoc-builder
 [Liferay Gradle Plugins Baseline]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-baseline
 [Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
+[Liferay Source Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/util/source-formatter
 [LPS-52675]: https://issues.liferay.com/browse/LPS-52675
 [LPS-58672]: https://issues.liferay.com/browse/LPS-58672
 [LPS-61099]: https://issues.liferay.com/browse/LPS-61099
@@ -742,6 +783,7 @@ moved to [Liferay Gradle Plugins].
 [LPS-68014]: https://issues.liferay.com/browse/LPS-68014
 [LPS-68131]: https://issues.liferay.com/browse/LPS-68131
 [LPS-68230]: https://issues.liferay.com/browse/LPS-68230
+[LPS-68289]: https://issues.liferay.com/browse/LPS-68289
 [LPS-68293]: https://issues.liferay.com/browse/LPS-68293
 [LPS-68297]: https://issues.liferay.com/browse/LPS-68297
 [LPS-68298]: https://issues.liferay.com/browse/LPS-68298
@@ -779,6 +821,10 @@ moved to [Liferay Gradle Plugins].
 [LPS-69470]: https://issues.liferay.com/browse/LPS-69470
 [LPS-69488]: https://issues.liferay.com/browse/LPS-69488
 [LPS-69492]: https://issues.liferay.com/browse/LPS-69492
+[LPS-69501]: https://issues.liferay.com/browse/LPS-69501
+[LPS-69518]: https://issues.liferay.com/browse/LPS-69518
+[LPS-69606]: https://issues.liferay.com/browse/LPS-69606
+[LPS-69618]: https://issues.liferay.com/browse/LPS-69618
 [LRDOCS-2594]: https://issues.liferay.com/browse/LRDOCS-2594
 [LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
 [LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981